requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
flask-mysql-connector==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==2.0.0
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.2.3
